% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_group_test.R
\name{mrs_group_test}
\alias{mrs_group_test}
\title{Statistical test for group separation of MRS}
\usage{
mrs_group_test(res)
}
\arguments{
\item{res}{Result object from \code{mrs_splitvalidate()}, containing \code{results_table} and phenotype.}
}
\value{
A named list with test method and p-value.
}
\description{
Performs Wilcoxon rank-sum test on MRS scores between case and control groups.
}
