% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrs_nested.R
\name{mrs_nested}
\alias{mrs_nested}
\title{Multi Resolution Scanning for nested design}
\usage{
mrs_nested(X, G, H, n_groups = length(unique(G)), n_subgroups = NULL,
  Omega = "default", K = 6, init_state = c(0.8, 0.2, 0), beta = 1,
  gamma = 0.3, eta = 0.3, alpha = 0.5, nu_vec = exp(seq(-1, 4)),
  return_global_null = TRUE, return_tree = TRUE)
}
\arguments{
\item{X}{Matrix of the data. Each row represents an observation.}

\item{G}{Numeric vector of the group label of each observation. Labels are integers starting from 1.}

\item{H}{Numeric vector of the replicate label of each observation. Labels are integers starting from 1.}

\item{n_groups}{Number of groups.}

\item{n_subgroups}{Vector indicating the number of replicates for each grop.}

\item{Omega}{Matrix defining the vertices of the sample space.
The \code{"default"} option defines a hyperrectangle containing all the data points.
Otherwise the user can define a matrix  where each row represents a dimension,  and the two columns contain the associated lower and upper limit.}

\item{K}{Depth of the tree. Default is \code{K = 5}, while the maximum is \code{K = 14}.}

\item{init_state}{Initial state of the hidden Markov process.
The three states are \emph{null}, \emph{altenrative} and \emph{prune}, respectively.}

\item{beta}{Spatial clustering parameter of the transition probability matrix.}

\item{gamma}{Parameter of the transition probability matrix.}

\item{alpha}{Pseudo-counts of the Beta random probability assignments.}

\item{nu_vec}{The support of the discrete uniform prior on nu.}

\item{return_global_null}{Boolean indicating whether to return the marginal posterior probability of the global null.}

\item{return_tree}{Boolean indicating whether to return the posterior representative tree.}
}
\value{
An \code{mrs} object.
}
\description{
This function executes the Multi Resolution Scanning algorithm to detect differences
across the distributions of multiple groups having multiple replicates.
}
\examples{
set.seed(1)
n = 20
p = 2
X = matrix(c(runif(p*n/2),rbeta(p*n/2, 1, 4)), nrow=n, byrow=TRUE)
G = c(rep(1,n/2), rep(2,n/2))
H = sample(3,2*n, replace = TRUE  )
ans = mrs_nested(X=X, G=G, H=H)
}

