% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.model2mdr.R
\name{parse_model_to_mdr}
\alias{parse_model_to_mdr}
\title{Parse MB-MDR .model File to Structured MDR Model List}
\usage{
parse_model_to_mdr(raw_model, row_per_pair = 14)
}
\arguments{
\item{raw_model}{A \code{data.frame} read from a \code{.model.txt} file.}

\item{row_per_pair}{Number of rows per SNP pair. Default is 14.}
}
\value{
A named list where each element is a SNP pair and includes:
\describe{
\item{affected}{3-row matrix for affected counts.}
\item{unaffected}{3-row matrix for unaffected counts.}
\item{HLO}{3×3 matrix of -1, 0, 1 encoding.}
}
}
\description{
Parses a raw \code{.model.txt} file into a structured list containing affected/unaffected counts and HLO matrices.
}
\examples{
\dontrun{
raw_model <- fread("single_model_2D.txt", fill = TRUE, header = FALSE)
mdr_model <- parse_model_to_mdr(raw_model)
}
}
