% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrs_score_dist.R
\name{plot_mrs_score_dist}
\alias{plot_mrs_score_dist}
\title{Plot risk score distribution by phenotype (test set)}
\arguments{
\item{res}{Result list from mrs_splitvalidate() with results_table.}
}
\value{
A ggplot violin plot annotated with AUC and Accuracy.
}
\description{
Violin + box plot showing score distribution across phenotypes.
}
