% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrs_validate.R
\name{plot_mrs_validate}
\alias{plot_mrs_validate}
\title{Plot validation results from mrs_splitvalidate}
\usage{
plot_mrs_validate(result, metric = "correlation", log_lambda = TRUE)
}
\arguments{
\item{result}{A result list from \code{mrs_splitvalidate()}.}

\item{metric}{One of \code{"correlation"} (default), \code{"auc"}, \code{"accuracy"}, \code{"f1"}.}

\item{log_lambda}{Logical, whether to log-transform x-axis.}
}
\value{
A base R plot of validation metric vs. lambda.
}
\description{
Plots lambda vs correlation.
}
